requirements.txt
langgraph==0.6.6
langchain==0.3.27
openai
python-dotenv
langchain_openai